Refactor CUPS package: don't treat CUPS API as thread-safe.
The CUPS API documentation claims that it is thread-safe, however
wild panics appear often when the connector handles multiple jobs.

Therefore, I moved all CUPS calls that use the http_t connection
to a new cupsCore class to isolate the connection and protect
it with a mutex.

I tried protecting the http_t connection with a RWMutex, using
the write lock for reconnect and the read lock for all else. Even
that isn't enough to fix the problem.

Change-Id: I9b7d35427fc51331a26584f868f1122a3c50a532