Ensure we don't forget about Color and Duplex options on restart
The CDD struct was overloaded with some local fields, and
these were getting lost during synchronization if we did a clean restart.
To ensure the diff/sync logic is simple, we don't overload CDD anymore.
Instead, we pack key:value pairs into CDD vendor fields when possible and
add fields to Printer when not.