Implement Privet on Windows fixes #246
First shot at Privet on Windows for local printing. Code needs a lot of
work but on my test machine, local printing seems to work for basic
scenarios.

This patch uses github.com/andrewtj/dnssd for mDNS which in turn relies
on Apple's Bonjour Service for Windows which must be installed:

https://support.apple.com/kb/dl999?locale=en_US

Apple does allow Bonjour Service to be distrubuted in some cases so we
may be able to wrap it into the MSI installer but for now it will need
to be installed before cloud-print-connector.