Handle mDNS collisions in privet with Avahi
Avahi makes it a bit tricky to handle service name collisions,
so it was not implemented before. Now multiple printers with
the same name can be registered on the same local network. This
comes up in the case where a privet-enabled printer is also
connected with the Cloud Print Connector.

Fixes #342.