Handle mDNS collisions in privet with Avahi (#359)
* Remove unused AvahiThreadedPoll parameters from several C functions

* Factor out populateGroup from addAvahiGroup

We will need to reuse the code in populateGroup to do resetAvahiGroup soon.

* Fix inconsistency in "serviceName" vs "service_name" in avahi.h

* Introduce resetAvahiGroup, which will be used for recovering from collision

* Add printer_name parameter to addAvahiGroup

When resolving collisions, the printer name will diverge from the
mDNS service name.

* Handle mDNS collisions in privet with Avahi

Avahi makes it a bit tricky to handle service name collisions,
so it was not implemented before. Now multiple printers with
the same name can be registered on the same local network. This
comes up in the case where a privet-enabled printer is also
connected with the Cloud Print Connector.

Fixes #342.